chore: move api tests
