+++ tests/api/core/strapi/document-service/delete.test.api.ts
