+++ tests/api/core/strapi/document-service/middlewares.test.api.ts
