+++ tests/api/core/strapi/document-service/resources/fixtures/article.json
