+++ tests/api/core/strapi/document-service/resources/fixtures/author.json
