+++ tests/api/core/strapi/document-service/resources/fixtures/category.json
