+++ tests/api/core/strapi/document-service/resources/fixtures/index.js
