+++ tests/api/core/strapi/document-service/resources/schemas/article.js
