+++ tests/api/core/strapi/document-service/resources/schemas/author.js
