+++ tests/api/core/strapi/document-service/resources/schemas/category.js
