+++ tests/api/core/strapi/document-service/resources/schemas/comp.js
