+++ tests/api/core/strapi/document-service/resources/schemas/dz-comp.js
