+++ tests/api/core/strapi/document-service/resources/schemas/dz-other-comp.js
