+++ tests/api/core/strapi/document-service/resources/schemas/index.js
