+++ tests/api/core/strapi/document-service/resources/types/components.d.ts
