+++ tests/api/core/strapi/document-service/resources/types/contentTypes.d.ts
