+++ tests/api/core/strapi/document-service/utils/index.ts
